+++ CHANGELOG.md
@@ -17,6 +17,16 @@
 ![aslistcmd.png](https://i.loli.net/2019/04/05/5ca74815a01a4.png)
 ![aslistcmd_win.png](https://i.loli.net/2019/04/05/5ca74819b276e.png)
 
+* 新增自定义命令 `aspowershell [on|off]`, 开启/关闭 PowerShell 模式
+
+> 如果`ascmd`命令指定的PowerShell解释器文件名中包函`powershell`关键字, 会自动启用 PowerShell 模式, 如下图:
+
+![aspowershell_default](https://i.loli.net/2019/04/05/5ca753673efe7.png)
+
+> 如果指定的 PowerShell 解释器文件名中不包含 `powershell` 关键字, 则需要手动使用该命令，启用 PowerShell 模式。如果关闭了 PowerShell 模式,则会执行出错，如下图:
+
+![aspowershell_switch](https://i.loli.net/2019/04/05/5ca75368d85fa.png)
+
 ### 其它
 
 * 修复默认设置保存时导致 bookmarks 清空的问题